[Refactoring] Move ReplaceBodiesWithFatalError to separate file
